+++ plugins/inputs/cloud_pubsub/pubsub_wrappers.go
@@ -1,4 +1,4 @@
-package google_pubsub
+package cloud_pubsub
 
 import (
 	"cloud.google.com/go/pubsub"
